Undo Changes

+++ controller/hostController.js
@@ -1,6 +1,15 @@
 const Home= require('../models/homes');
 const fs= require('fs');
 
+const randomString= (length) => {
+    const characters= '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
+    let result= '';
+    for(let i=0; i<length; i++){
+        result+= characters.charAt(Math.floor(Math.random() * characters.length));
+    }
+    return result;
+}
+
 exports.getAddHome= (req, res, next) => {
     res.render('host/edit-home', {
         pageTitle: 'Add home to airbnb', 
@@ -50,9 +59,40 @@ exports.postAddHome= (req, res, next) => {
     if(!photoFile) {
         return res.status(400).send("No image Uploaded");
     }
-    const home= new Home({houseName, price, location, rating, photo: photoFile.path, description, rules: rulesFile ? rulesFile.path : ''});
+    // const home= new Home({houseName, price, location, rating, photo: photoFile.path, description, rules: rulesFile ? rulesFile.path : ''});
+    // const normalizedPhotoPath = photoFile.path.replace(/\\/g, '/');
+    // const normalizedRulesPath = rulesFile ? rulesFile.path.replace(/\\/g, '/') : '';
+    // const home= new Home({houseName, price, location, rating, photo: normalizedPhotoPath, description, rules: normalizedRulesPath});
+    const photoName = randomString(10) + '-' + photoFile.originalname;
+    const normalizedPhotoPath = 'uploads/' + photoName;
+    
+    let normalizedRulesPath = '';
+    let rulesBuffer = null;
+    let rulesMimeType = '';
+    if (rulesFile) {
+        const rulesName = `rules-${randomString(10)}.pdf`;
+        normalizedRulesPath = 'rules/' + rulesName;
+        rulesBuffer = rulesFile.buffer;
+        rulesMimeType = rulesFile.mimetype;
+    }
+
+    const home = new Home({
+        houseName,
+        price,
+        location,
+        rating,
+        photo: normalizedPhotoPath,
+        photoBuffer: photoFile.buffer,
+        photoMimeType: photoFile.mimetype,
+        description,
+        rules: normalizedRulesPath,
+        rulesBuffer,
+        rulesMimeType
+    });
     home.save().then(() => {
         console.log("Home saved successfully");
+    }).catch(err => {
+        console.log("Error saving home", err);
     });
     res.redirect('/host/host-home-list');
 }
@@ -62,35 +102,64 @@ exports.postEditHome= (req, res, next) => {
     const photoFile = req.files && req.files['photo'] ? req.files['photo'][0] : null;
     const rulesFile = req.files && req.files['rules'] ? req.files['rules'][0] : null;
     Home.findById(id).then((home) => {
-        home.houseName= houseName;
-        home.price= price;
-        home.location= location;
-        home.rating= rating;
-        if(photoFile) {  // If user has uploaded a new image, then update the photo path...
-            fs.unlink(home.photo, (err) => {
-                if (err) {
-                    console.log("Error while deleting old photo", err);
-                }
-            });
-            home.photo= photoFile.path;
+        // home.houseName= houseName;
+        // home.price= price;
+        // home.location= location;
+        // home.rating= rating;
+        // if(photoFile) {  // If user has uploaded a new image, then update the photo path...
+        if (!home) {
+            console.log("Home not found for editing");
+            return res.redirect('/host/host-home-list');
         }
-        home.description= description;
-        if(rulesFile) {  // If user has uploaded a new rules file, then update the rules path...
-            fs.unlink(home.rules, (err) => {
-                if (err) {
-                    console.log("Error while deleting old rule files", err);
-                }
-            });
-            home.rules= rulesFile.path;
+        home.houseName = houseName;
+        home.price = price;
+        home.location = location;
+        home.rating = rating;
+        home.description = description;
+
+        if (photoFile) {
+            if (home.photo && fs.existsSync(home.photo)) {
+                fs.unlink(home.photo, (err) => {
+                    if (err) {
+                        console.log("Error while deleting old photo", err);
+                    }
+                });
+            // home.photo= photoFile.path;
+        //     home.photo= photoFile.path.replace(/\\/g, '/');
+        // }
+        // home.description= description;
+        // if(rulesFile) {  // If user has uploaded a new rules file, then update the rules path...
+            }
+            const photoName = randomString(10) + '-' + photoFile.originalname;
+            home.photo = 'uploads/' + photoName;
+            home.photoBuffer = photoFile.buffer;
+            home.photoMimeType = photoFile.mimetype;
+        }
+
+        if (rulesFile) {
+            if (home.rules && fs.existsSync(home.rules)) {
+                fs.unlink(home.rules, (err) => {
+                    if (err) {
+                        console.log("Error while deleting old rule files", err);
+                    }
+                });
+            // home.rules= rulesFile.path;
+            // home.rules= rulesFile.path.replace(/\\/g, '/');
+            }
+            const rulesName = `rules-${randomString(10)}.pdf`;
+            home.rules = 'rules/' + rulesName;
+            home.rulesBuffer = rulesFile.buffer;
+            home.rulesMimeType = rulesFile.mimetype;
         }
         home.save().then(result => {
             console.log("Home updated", result);
         }).catch(err => {
             console.log('Error while updating', err);
-        })
+        });
         res.redirect('/host/host-home-list');
     }).catch(err => {
         console.log('Error while finding home', err);
+        res.redirect('/host/host-home-list');
     });
 }
